agregando avatar a navBar
