--- client/src/pages/Avatar/index.jsx
@@ -1,33 +1,0 @@
-import Layout from "../../Layout";
-import { Link } from "react-router-dom";
-const Avatar = () => {
-  return (
-    <>
-      {/* <div className="flex items-center">
-        <div className="relative w-12 h-12 rounded-full overflow-hidden bg-gray-300">
-          <img
-            src=""
-            alt="User Avatar"
-            className="absolute inset-0 w-full h-full object-cover"
-          />
-          <span className="absolute inset-0 flex items-center justify-center text-white text-2xl font-bold text-shadow"></span>
-        </div>
-        <div className="ml-4">
-          <div className="text-lg font-medium">Bienvenido</div>
-          <div className="text-gray-500" id="login-status">
-            <Link to="/login"></Link>
-            <Avatar />
-          </div>
-        </div>
-        <button
-          className="ml-4 py-2 px-4 bg-blue-500 text-white rounded hover:bg-blue-600 hidden"
-          id="edit-profile-button"
-        >
-          Editar perfil
-        </button>
-      </div> */}
-    </>
-  );
-};
-
-export default Avatar;
